deployment_package/requirements.txt
# RemoteHive Backend API Dependencies
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
# Legacy SQL dependencies (needed for migration compatibility)
sqlalchemy==1.4.53
alembic==1.11.1
# pymysql==1.1.0  # For MySQL support
# psycopg2==2.9.9  # For PostgreSQL support (production) - commented out for local development

# MongoDB dependencies
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver
beanie==1.24.0  # Async ODM for MongoDB
dnspython==2.4.2  # Required for MongoDB SRV connections

# Authentication and security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Data validation and settings
pydantic==2.9.2
pydantic-settings==2.6.1

# Background tasks and caching
celery==5.3.4
redis==5.0.1
flower==2.0.1

# HTTP requests and utilities
requests==2.31.0
httpx==0.24.1

# Web scraping and browser automation
playwright==1.40.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Configuration and environment
python-dotenv==1.0.0

# Logging and monitoring
loguru==0.7.2

# Email templates and rendering
jinja2==3.1.2

# File handling and utilities
aiofiles==23.2.1
aiohttp==3.9.1
python-dateutil==2.8.2
pillow==10.4.0

# Development and testing
faker==20.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Web scraping and automation
scrapy==2.11.0
lxml==5.3.0
user-agent==0.1.10
fake-useragent==1.4.0
feedparser==6.0.11

# Anti-detection and proxy support
requests-html==0.10.0
cloudscraper==1.2.71
undetected-chromedriver==3.5.4

# CLI management tools
click==8.1.7
tabulate==0.9.0
rich==13.7.0

# Machine Learning and Data Science
scikit-learn==1.3.2
spacy==3.7.2
pandas==2.1.4
numpy==1.24.4
nltk==3.8.1
joblib==1.3.2
google-generativeai==0.3.2

# Additional utilities for production
gunicorn==21.2.0
psutil==5.9.6
